requirements.txt
# Core
python-dotenv==1.2.2

# Backend
fastapi==0.135.3
uvicorn==0.42.0
pydantic[email]==2.12.5

# Database
supabase

# Audio / Video
yt-dlp
requests

# Speech-to-Text
assemblyai

# AI / LLM
google-genai==1.73.0

# Frontend
streamlit

# NLP
spacy==3.8.14
